Refactored old ExportNodeCommand class
