+++ ApsimNG/ApsimNG.csproj
@@ -198,9 +198,9 @@
     <Compile Include="Commands\AddModelCommand.cs" />
     <Compile Include="Commands\BibTeX.cs" />
     <Compile Include="Commands\ChangeProperty.cs" />
-    <Compile Include="Commands\CreateModelDescriptionDoc.cs" />
+    <Compile Include="Commands\CreateModelDescriptionDocCommand.cs" />
     <Compile Include="Commands\DeleteModelCommand.cs" />
-    <Compile Include="Commands\ExportNodeCommand.cs" />
+    <Compile Include="Commands\CreateDocCommand.cs" />
     <Compile Include="Commands\IExplorerPresenter.cs" />
     <Compile Include="Commands\MoveModelCommand.cs" />
     <Compile Include="Commands\MoveModelUpDownCommand.cs" />
